feat: service-break-p04-extract-workers-service-001 - Run script pollers through the workers service
Move script poller supervision, command setup, output parsing, and backoff
handling into pkg/workers/service while root pkg/service delegates through
the new collaborator.
