+++ pkg/service/poller_watcher.go
@@ -1,12 +1,9 @@
 package service
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"path/filepath"
 	"strings"
 	"sync"
 	"time"
@@ -14,19 +11,12 @@ import (
 	"github.com/go-co-op/gocron/v2"
 	"github.com/jonboulle/clockwork"
 	"github.com/portpowered/infinite-you/pkg/config"
-	factory_context "github.com/portpowered/infinite-you/pkg/factory/context"
-	"github.com/portpowered/infinite-you/pkg/factory/requests"
 	"github.com/portpowered/infinite-you/pkg/hostedworkers"
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 	"github.com/portpowered/infinite-you/pkg/timework"
 	"github.com/portpowered/infinite-you/pkg/workers"
-	workerprompting "github.com/portpowered/infinite-you/pkg/workers/prompting"
+	workersservice "github.com/portpowered/infinite-you/pkg/workers/service"
 	"go.uber.org/zap"
-)
-
-const (
-	pollerRestartBackoffMin = 25 * time.Millisecond
-	pollerRestartBackoffMax = 250 * time.Millisecond
 )
 
 func (fs *FactoryService) startPollerWatchersForRuntime(
@@ -66,11 +56,7 @@ func (fs *FactoryService) startPollerWatchersForRuntime(
 		}
 		switch {
 		case interfaces.IsScriptWorkerType(workerDef.Type):
-			sidecars.Add(1)
-			go func() {
-				defer sidecars.Done()
-				fs.superviseScriptPoller(ctx, runtimeCfg, ws, workerDef, submitter)
-			}()
+			fs.workersSchedulerService().StartScriptPoller(ctx, sidecars, runtimeCfg, ws, workerDef, workersservice.WorkRequestSubmitter(submitter))
 		case interfaces.IsPollerWorkerType(workerDef.Type):
 			if workerDef.Provider != interfaces.HostedWorkerProviderLinear {
 				fs.logger.Warn("hosted poller disabled",
@@ -88,359 +74,26 @@ func (fs *FactoryService) startPollerWatchersForRuntime(
 	}
 }
 
-func (fs *FactoryService) superviseScriptPoller(
-	ctx context.Context,
-	runtimeCfg interfaces.RuntimeConfigLookup,
-	workstation interfaces.FactoryWorkstationConfig,
-	workerDef *interfaces.WorkerConfig,
-	submitter workRequestSubmitter,
-) {
-	logger := fs.pollerLogger(workstation, workerDef)
-	runner := fs.pollerCommandRunner()
-	backoffClock := fs.pollerSupervisorClock()
-	attempt := 0
-	logger.Info("script poller started")
-	defer func() {
-		logger.Info("script poller stopped", zap.String("reason", pollerStopReason(ctx.Err())))
-	}()
-
-	for {
-		if ctx.Err() != nil {
-			return
-		}
-
-		attempt++
-		runErr := fs.runScriptPoller(ctx, runner, runtimeCfg, workstation, workerDef, submitter)
-		if ctx.Err() != nil {
-			return
-		}
-
-		backoff := pollerRestartBackoff(attempt)
-		logger.Warn("script poller restarting",
-			zap.Int("attempt", attempt),
-			zap.Duration("backoff", backoff),
-			zap.Error(runErr),
-		)
-
-		select {
-		case <-ctx.Done():
-			return
-		case <-backoffClock.After(backoff):
-		}
-	}
-}
-
-func (fs *FactoryService) runScriptPoller(
-	ctx context.Context,
-	runner workers.CommandRunner,
-	runtimeCfg interfaces.RuntimeConfigLookup,
-	workstation interfaces.FactoryWorkstationConfig,
-	workerDef *interfaces.WorkerConfig,
-	submitter workRequestSubmitter,
-) error {
-	commandReq, err := scriptPollerCommandRequest(runtimeCfg, workstation, workerDef)
-	if err != nil {
-		return err
-	}
-
-	execCtx := ctx
-	timeout, err := scriptPollerExecutionTimeout(workstation, workerDef)
-	if err != nil {
-		return err
-	}
-	if timeout > 0 {
-		var cancel context.CancelFunc
-		execCtx, cancel = context.WithTimeout(ctx, timeout)
-		defer cancel()
-	}
-
-	result, runErr := runner.Run(execCtx, commandReq)
-	if ctx.Err() != nil {
-		return ctx.Err()
-	}
-	if runErr != nil {
-		if errors.Is(runErr, context.DeadlineExceeded) {
-			return fmt.Errorf("script poller timed out after %s", timeout)
-		}
-		return fmt.Errorf("script poller execution failed: %w", runErr)
-	}
-	if result.ExitCode != 0 {
-		return fmt.Errorf("script poller exited with code %d", result.ExitCode)
-	}
-	request, hasOutput, err := parseScriptPollerOutput(result.Stdout)
-	if err != nil {
-		return err
-	}
-	if hasOutput {
-		if submitter == nil {
-			return fmt.Errorf("script poller submitter is not available")
-		}
-		if err := submitter(ctx, request); err != nil {
-			return fmt.Errorf("script poller submit failed: %w", err)
-		}
-	}
-	return fmt.Errorf("script poller exited unexpectedly")
-}
-
-func (fs *FactoryService) pollerCommandRunner() workers.CommandRunner {
-	if fs != nil && fs.coordinatorPolicy().commandRunnerOverride != nil {
-		return fs.coordinatorPolicy().commandRunnerOverride
-	}
-	return workers.ExecCommandRunner{}
-}
-
-func (fs *FactoryService) pollerSupervisorClock() clockwork.Clock {
+func (fs *FactoryService) workersSchedulerService() *workersservice.Service {
+	clock := clockwork.NewRealClock()
 	if fs != nil {
 		if supervisorClock, ok := fs.clock.(clockwork.Clock); ok && supervisorClock != nil {
-			return supervisorClock
+			clock = supervisorClock
 		}
 	}
-	return clockwork.NewRealClock()
-}
-
-func (fs *FactoryService) pollerLogger(workstation interfaces.FactoryWorkstationConfig, workerDef *interfaces.WorkerConfig) *zap.Logger {
-	if fs == nil || fs.logger == nil {
-		return zap.NewNop()
+	var runner workers.CommandRunner = workers.ExecCommandRunner{}
+	if fs != nil && fs.coordinatorPolicy().commandRunnerOverride != nil {
+		runner = fs.coordinatorPolicy().commandRunnerOverride
 	}
-	return fs.logger.With(
-		zap.String("workstation", workstation.Name),
-		zap.String("worker", workerDef.Name),
-	)
-}
-
-func pollerRestartBackoff(attempt int) time.Duration {
-	if attempt <= 1 {
-		return pollerRestartBackoffMin
+	logger := zap.NewNop()
+	if fs != nil && fs.logger != nil {
+		logger = fs.logger
 	}
-	backoff := pollerRestartBackoffMin
-	for i := 1; i < attempt && backoff < pollerRestartBackoffMax; i++ {
-		backoff *= 2
-		if backoff >= pollerRestartBackoffMax {
-			return pollerRestartBackoffMax
-		}
-	}
-	return backoff
-}
-
-func pollerStopReason(err error) string {
-	switch {
-	case errors.Is(err, context.Canceled):
-		return "context canceled"
-	case errors.Is(err, context.DeadlineExceeded):
-		return "deadline exceeded"
-	case err != nil:
-		return err.Error()
-	default:
-		return "completed"
-	}
-}
-
-func scriptPollerCommandRequest(
-	runtimeCfg interfaces.RuntimeConfigLookup,
-	workstation interfaces.FactoryWorkstationConfig,
-	workerDef *interfaces.WorkerConfig,
-) (workers.CommandRequest, error) {
-	if runtimeCfg == nil {
-		return workers.CommandRequest{}, fmt.Errorf("runtime config is required")
-	}
-	if workerDef == nil {
-		return workers.CommandRequest{}, fmt.Errorf("script poller worker is required")
-	}
-	if strings.TrimSpace(workerDef.Command) == "" {
-		return workers.CommandRequest{}, fmt.Errorf("script poller worker %q is missing command", workerDef.Name)
-	}
-
-	requestContext := &factory_context.FactoryContext{}
-	if resolved, err := workerprompting.ResolveTemplateFields(
-		workstation.WorkingDirectory,
-		workstation.Env,
-		nil,
-		requestContext,
-		workstation.Worktree,
-	); err != nil {
-		return workers.CommandRequest{}, fmt.Errorf("resolve poller workstation fields: %w", err)
-	} else if resolved != nil {
-		requestContext.WorkDirectory = resolved.WorkingDirectory
-		requestContext.EnvVars = resolved.Env
-		if requestContext.WorkDirectory == "" {
-			requestContext.WorkDirectory = resolved.Worktree
-		}
-	}
-
-	workDir := requestContext.WorkDirectory
-	if workDir != "" && !filepath.IsAbs(workDir) {
-		baseDir := runtimeCfg.RuntimeBaseDir()
-		if baseDir == "" {
-			baseDir = runtimeCfg.FactoryDir()
-		}
-		if baseDir != "" {
-			workDir = filepath.Clean(filepath.Join(baseDir, filepath.FromSlash(workDir)))
-		}
-	}
-	if workDir == "" {
-		workDir = pollerRuntimeWorkingDirectory(runtimeCfg)
-	}
-
-	req := workers.CommandRequest{
-		Command:         resolvePortableFactoryScriptReference(runtimeCfg.FactoryDir(), workerDef.Command),
-		Args:            resolvePortableFactoryScriptReferences(runtimeCfg.FactoryDir(), workerDef.Args),
-		Env:             commandEnvWithResolvedVars(requestContext.EnvVars),
-		WorkDir:         workDir,
-		WorkerType:      workerDef.Name,
-		WorkstationName: workstation.Name,
-	}
-	return req, nil
-}
-
-func pollerRuntimeWorkingDirectory(runtimeCfg interfaces.RuntimeConfigLookup) string {
-	if runtimeCfg == nil {
-		return ""
-	}
-	baseDir := strings.TrimSpace(runtimeCfg.RuntimeBaseDir())
-	if baseDir == "" {
-		baseDir = strings.TrimSpace(runtimeCfg.FactoryDir())
-	}
-	if baseDir == "" {
-		return ""
-	}
-	return filepath.Clean(baseDir)
-}
-
-func scriptPollerExecutionTimeout(workstation interfaces.FactoryWorkstationConfig, workerDef *interfaces.WorkerConfig) (time.Duration, error) {
-	timeout, err := config.WorkstationExecutionTimeout(&workstation)
-	if err != nil {
-		return 0, err
-	}
-	if timeout > 0 {
-		return timeout, nil
-	}
-	if workerDef != nil && strings.TrimSpace(workerDef.Timeout) != "" {
-		parsed, err := time.ParseDuration(workerDef.Timeout)
-		if err != nil {
-			return 0, fmt.Errorf("invalid worker timeout %q: %w", workerDef.Timeout, err)
-		}
-		if parsed > 0 {
-			return parsed, nil
-		}
-	}
-	return 0, nil
-}
-
-func parseScriptPollerOutput(stdout []byte) (interfaces.WorkRequest, bool, error) {
-	trimmed := bytes.TrimSpace(stdout)
-	if len(trimmed) == 0 {
-		return interfaces.WorkRequest{}, false, nil
-	}
-
-	var envelope scriptPollerOutputEnvelope
-	if err := json.Unmarshal(trimmed, &envelope); err != nil {
-		return interfaces.WorkRequest{}, true, fmt.Errorf("script poller emitted malformed stdout: %w", err)
-	}
-	if len(envelope.Events) > 0 {
-		return interfaces.WorkRequest{}, true, fmt.Errorf("script poller emitted unsupported raw factory events")
-	}
-	if len(envelope.Request) > 0 && len(envelope.Submissions) > 0 {
-		return interfaces.WorkRequest{}, true, fmt.Errorf("script poller stdout must contain either request or submissions, not both")
-	}
-	if len(envelope.Request) > 0 {
-		request, err := requests.ParseCanonicalWorkRequestJSON(envelope.Request)
-		if err != nil {
-			return interfaces.WorkRequest{}, true, fmt.Errorf("script poller emitted malformed stdout: %w", err)
-		}
-		if err := validateScriptPollerWorkRequest(request); err != nil {
-			return interfaces.WorkRequest{}, true, err
-		}
-		return request, true, nil
-	}
-	if len(envelope.Submissions) > 0 {
-		request, err := scriptPollerWorkRequestFromSubmissions(envelope.Submissions)
-		if err != nil {
-			return interfaces.WorkRequest{}, true, err
-		}
-		return request, true, nil
-	}
-
-	request, err := requests.ParseCanonicalWorkRequestJSON(trimmed)
-	if err != nil {
-		return interfaces.WorkRequest{}, true, fmt.Errorf("script poller emitted malformed stdout: %w", err)
-	}
-	if err := validateScriptPollerWorkRequest(request); err != nil {
-		return interfaces.WorkRequest{}, true, err
-	}
-	return request, true, nil
-}
-
-type scriptPollerOutputEnvelope struct {
-	Request     json.RawMessage `json:"request"`
-	Submissions json.RawMessage `json:"submissions"`
-	Events      json.RawMessage `json:"events"`
-}
-
-func scriptPollerWorkRequestFromSubmissions(data []byte) (interfaces.WorkRequest, error) {
-	var submissions []interfaces.SubmitRequest
-	if err := json.Unmarshal(data, &submissions); err != nil {
-		return interfaces.WorkRequest{}, fmt.Errorf("script poller emitted malformed stdout: decode submissions: %w", err)
-	}
-	if len(submissions) == 0 {
-		return interfaces.WorkRequest{}, fmt.Errorf("script poller emitted malformed stdout: submissions must contain at least one item")
-	}
-
-	request := requests.WorkRequestFromSubmitRequests(submissions)
-	if strings.TrimSpace(request.RequestID) == "" {
-		return interfaces.WorkRequest{}, fmt.Errorf("script poller emitted malformed stdout: submissions must share a non-empty requestId")
-	}
-	return request, nil
-}
-
-func validateScriptPollerWorkRequest(request interfaces.WorkRequest) error {
-	if request.Type != interfaces.WorkRequestTypeFactoryRequestBatch {
-		return fmt.Errorf("script poller emitted malformed stdout: unsupported work request type %q", request.Type)
-	}
-	if strings.TrimSpace(request.RequestID) == "" {
-		return fmt.Errorf("script poller emitted malformed stdout: work request must set requestId")
-	}
-	return nil
-}
-
-func commandEnvWithResolvedVars(vars map[string]string) []string {
-	if len(vars) == 0 {
-		return nil
-	}
-
-	env := make([]string, 0, len(vars))
-	for key, value := range vars {
-		env = append(env, key+"="+value)
-	}
-	return env
-}
-
-func resolvePortableFactoryScriptReferences(factoryDir string, args []string) []string {
-	if len(args) == 0 {
-		return nil
-	}
-	resolved := make([]string, len(args))
-	for i, arg := range args {
-		resolved[i] = resolvePortableFactoryScriptReference(factoryDir, arg)
-	}
-	return resolved
-}
-
-func resolvePortableFactoryScriptReference(factoryDir, raw string) string {
-	if strings.TrimSpace(factoryDir) == "" {
-		return raw
-	}
-
-	trimmed := strings.TrimSpace(raw)
-	normalized := filepath.ToSlash(trimmed)
-	if !strings.HasPrefix(normalized, "factory/scripts/") {
-		return raw
-	}
-
-	relativePath := strings.TrimPrefix(normalized, "factory/scripts/")
-	if relativePath == "" {
-		return raw
-	}
-	return filepath.Join(factoryDir, "scripts", filepath.FromSlash(relativePath))
+	return workersservice.New(workersservice.Config{
+		Logger:        logger,
+		Clock:         clock,
+		CommandRunner: runner,
+	})
 }
 
 const (
