+++ pkg/service/poller_watcher_test.go
@@ -23,6 +23,7 @@ import (
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 	"github.com/portpowered/infinite-you/pkg/petri"
 	"github.com/portpowered/infinite-you/pkg/workers"
+	workersservice "github.com/portpowered/infinite-you/pkg/workers/service"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zaptest/observer"
 )
@@ -32,138 +33,6 @@ const (
 	canonicalScriptPollerWorkerName      = "poller-script"
 	canonicalScriptPollerCommand         = "factory/scripts/poller.sh"
 )
-
-func TestRunScriptPoller_SubmitsCanonicalWorkRequestStdoutToFactoryService(t *testing.T) {
-	factoryDir := t.TempDir()
-	workRequestJSON := []byte(`{
-		"requestId":"linear-issue-batch-1",
-		"type":"FACTORY_REQUEST_BATCH",
-		"works":[{"name":"issue-123","workTypeName":"task","payload":{"id":"ISSUE-123"}}]
-	}`)
-	runner := &pollerSequenceCommandRunner{
-		outcomes: []pollerRunOutcome{{result: workers.CommandResult{Stdout: workRequestJSON}}},
-	}
-	submitted := &aggregateSnapshotFactory{}
-	svc := &FactoryService{
-		policy: serviceCoordinatorPolicyFromConfig(&FactoryServiceConfig{CommandRunnerOverride: runner}),
-		logger: zap.NewNop(),
-	}
-	poller := newCanonicalScriptPollerWorkstation()
-	worker := newCanonicalScriptPollerWorker()
-	runtimeCfg := newLoadedFactoryConfigForServiceTest(
-		t,
-		factoryDir,
-		&interfaces.FactoryConfig{Workers: []interfaces.WorkerConfig{{Name: canonicalScriptPollerWorkerName}}, Workstations: []interfaces.FactoryWorkstationConfig{poller}},
-		map[string]*interfaces.WorkerConfig{canonicalScriptPollerWorkerName: worker},
-		map[string]*interfaces.FactoryWorkstationConfig{poller.Name: &poller},
-	)
-
-	err := svc.runScriptPoller(
-		context.Background(),
-		runner,
-		runtimeCfg,
-		poller,
-		worker,
-		submitWorkRequestWithFactory(submitted),
-	)
-	if err == nil || !strings.Contains(err.Error(), "exited unexpectedly") {
-		t.Fatalf("runScriptPoller error = %v, want unexpected exit after successful submit", err)
-	}
-	if submitted.submitCalls != 1 {
-		t.Fatalf("submit calls = %d, want 1", submitted.submitCalls)
-	}
-	if len(submitted.submissions) != 1 {
-		t.Fatalf("submitted requests = %d, want 1", len(submitted.submissions))
-	}
-	if submitted.submissions[0].RequestID != "linear-issue-batch-1" {
-		t.Fatalf("submitted request ID = %q, want linear-issue-batch-1", submitted.submissions[0].RequestID)
-	}
-	if submitted.submissions[0].Type != interfaces.WorkRequestTypeFactoryRequestBatch {
-		t.Fatalf("submitted request type = %q, want FACTORY_REQUEST_BATCH", submitted.submissions[0].Type)
-	}
-}
-
-func TestRunScriptPoller_SubmitsSubmitStyleRecordsStdoutToFactoryService(t *testing.T) {
-	factoryDir := t.TempDir()
-	envelopeJSON := []byte(`{
-		"submissions":[
-			{
-				"requestId":"linear-issue-batch-2",
-				"workId":"linear-issue-124",
-				"name":"issue-124",
-				"workTypeName":"task",
-				"traceId":"trace-124"
-			}
-		]
-	}`)
-	runner := &pollerSequenceCommandRunner{
-		outcomes: []pollerRunOutcome{{result: workers.CommandResult{Stdout: envelopeJSON}}},
-	}
-	submitted := &aggregateSnapshotFactory{}
-	svc := &FactoryService{
-		policy: serviceCoordinatorPolicyFromConfig(&FactoryServiceConfig{CommandRunnerOverride: runner}),
-		logger: zap.NewNop(),
-	}
-	poller := newCanonicalScriptPollerWorkstation()
-	worker := newCanonicalScriptPollerWorker()
-	runtimeCfg := newLoadedFactoryConfigForServiceTest(
-		t,
-		factoryDir,
-		&interfaces.FactoryConfig{Workers: []interfaces.WorkerConfig{{Name: canonicalScriptPollerWorkerName}}, Workstations: []interfaces.FactoryWorkstationConfig{poller}},
-		map[string]*interfaces.WorkerConfig{canonicalScriptPollerWorkerName: worker},
-		map[string]*interfaces.FactoryWorkstationConfig{poller.Name: &poller},
-	)
-
-	err := svc.runScriptPoller(
-		context.Background(),
-		runner,
-		runtimeCfg,
-		poller,
-		worker,
-		submitWorkRequestWithFactory(submitted),
-	)
-	if err == nil || !strings.Contains(err.Error(), "exited unexpectedly") {
-		t.Fatalf("runScriptPoller error = %v, want unexpected exit after successful submit", err)
-	}
-	if len(submitted.submissions) != 1 {
-		t.Fatalf("submitted requests = %d, want 1", len(submitted.submissions))
-	}
-	workRequest := submitted.submissions[0]
-	if workRequest.RequestID != "linear-issue-batch-2" {
-		t.Fatalf("submitted request ID = %q, want linear-issue-batch-2", workRequest.RequestID)
-	}
-	if workRequest.Works == nil || len(workRequest.Works) != 1 {
-		t.Fatalf("submitted works = %#v, want one canonical work item", workRequest.Works)
-	}
-	if workRequest.Works[0].WorkID != "linear-issue-124" {
-		t.Fatalf("submitted work ID = %q, want linear-issue-124", workRequest.Works[0].WorkID)
-	}
-}
-
-func TestScriptPollerCommandRequest_DefaultsEmptyWorkingDirectoryToRuntimeBaseDirectory(t *testing.T) {
-	factoryDir := t.TempDir()
-	runtimeBaseDir := t.TempDir()
-	runtimeCfg := newScriptPollerLoadedRuntimeConfigForServiceTest(
-		t,
-		factoryDir,
-		scriptPollerRuntimeConfigOptions{
-			poller: newCanonicalScriptPollerWorkstation(),
-		},
-	)
-	runtimeCfg.SetRuntimeBaseDir(runtimeBaseDir)
-
-	req, err := scriptPollerCommandRequest(
-		runtimeCfg,
-		newCanonicalScriptPollerWorkstation(),
-		newCanonicalScriptPollerWorker("--mode", "watch"),
-	)
-	if err != nil {
-		t.Fatalf("scriptPollerCommandRequest: %v", err)
-	}
-	if req.WorkDir != runtimeBaseDir {
-		t.Fatalf("poller workdir = %q, want %q", req.WorkDir, runtimeBaseDir)
-	}
-}
 
 func TestFactoryService_StartLiveRuntimeSidecars_StartsScriptPollerForPollerRunWorkstationType(t *testing.T) {
 	start := time.Date(2026, time.June, 16, 9, 0, 0, 0, time.UTC)
@@ -267,7 +136,7 @@ func TestFactoryService_StartLiveRuntimeSidecars_StartsOnlyScriptPollersAndResta
 
 	waitForPollerRunnerCalls(t, runner, 1, time.Second)
 	waitForFakeClockWaiters(t, fakeClock, 1)
-	fakeClock.Advance(pollerRestartBackoffMin)
+	fakeClock.Advance(workersservice.ScriptPollerRestartBackoffMin)
 	waitForPollerRunnerCalls(t, runner, 2, time.Second)
 
 	reqs := runner.requests()
@@ -645,7 +514,7 @@ func TestFactoryService_StartLiveRuntimeSidecars_RestartsScriptPollerOnMalformed
 
 	waitForPollerRunnerCalls(t, runner, 1, time.Second)
 	waitForFakeClockWaiters(t, fakeClock, 1)
-	fakeClock.Advance(pollerRestartBackoffMin)
+	fakeClock.Advance(workersservice.ScriptPollerRestartBackoffMin)
 	waitForPollerRunnerCalls(t, runner, 2, time.Second)
 
 	if observedLogs.FilterMessage("script poller restarting").Len() == 0 {
@@ -858,25 +727,6 @@ func newScriptPollerRuntimeHandleForWorkstation(
 				},
 			),
 		},
-	}
-}
-
-func TestParseScriptPollerOutput_RejectsUnsupportedRawFactoryEvents(t *testing.T) {
-	rawEventJSON, err := json.Marshal(map[string]any{
-		"events": []map[string]any{{
-			"type": "WORK_REQUEST",
-		}},
-	})
-	if err != nil {
-		t.Fatalf("marshal raw event payload: %v", err)
-	}
-
-	_, hasOutput, parseErr := parseScriptPollerOutput(rawEventJSON)
-	if !hasOutput {
-		t.Fatal("expected raw event payload to count as poller output")
-	}
-	if parseErr == nil || !strings.Contains(parseErr.Error(), "unsupported raw factory events") {
-		t.Fatalf("parse error = %v, want unsupported raw factory events", parseErr)
 	}
 }
 
@@ -1297,19 +1147,6 @@ func waitForObservedLogMessage(t *testing.T, logs *observer.ObservedLogs, messag
 		time.Sleep(10 * time.Millisecond)
 	}
 	t.Fatalf("timed out waiting for log message %q", message)
-}
-
-func TestParseScriptPollerOutput_RejectsMalformedStdout(t *testing.T) {
-	_, hasOutput, err := parseScriptPollerOutput([]byte("submitted work\n"))
-	if !hasOutput {
-		t.Fatal("expected non-empty stdout to count as poller output")
-	}
-	if err == nil {
-		t.Fatal("expected malformed stdout error")
-	}
-	if !strings.Contains(err.Error(), "malformed stdout") {
-		t.Fatalf("error = %v, want malformed stdout", err)
-	}
 }
 
 func TestFactoryService_RequiredInputCronKeepsTimeWorkPendingWhenInputMissing(t *testing.T) {
